+++ src/Lumina/Excel/MismatchedColumnHashException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Lumina.Excel;
+
+/// <summary>Exception indicating that the requested row type's column hash is different from game data.</summary>
+/// <inheritdoc/>
+public sealed class MismatchedColumnHashException( uint typeHash, uint gameHash, string? paramName ) :
+    ArgumentException(
+        $"The requested row type has a column hash that is different from game data. (Type: {typeHash:X08}, Game: {gameHash:X08})",
+        paramName )
+{ }
